Rename DevP2P to P2P for convenient
