+++ lib/ciri/p2p/dial_scheduler.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+
 # Copyright (c) 2018 by Jiang Jinyang <jjyruby@gmail.com>
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -21,29 +22,47 @@
 # THE SOFTWARE.
 
 
+require 'async'
+require 'ciri/utils/logger'
+
 module Ciri
-  module DevP2P
-    module RLPX
+  module P2P
 
-      # class used to store rplx protocol secrets
-      class Secrets
-        attr_reader :remote_id, :aes, :mac
-        attr_accessor :egress_mac, :ingress_mac
+    # DialScheduler
+    # establish outoging connections
+    class DialScheduler
+      include Utils::Logger
 
-        def initialize(remote_id: nil, aes:, mac:)
-          @remote_id = remote_id
-          @aes = aes
-          @mac = mac
-        end
+      MAX_ACTIVE_DIAL_TASKS = 16
 
-        def ==(other)
-          self.class == other.class &&
-            remote_id == other.remote &&
-            aes == other.aes &&
-            mac == other.mac
+      def initialize(network_state, dialer, discovery_service)
+        @network_state = network_state
+        @running_dialing = 0
+        @dialer = dialer
+        @discovery_service = discovery_service
+      end
+
+      def run(task: Async::Task.current)
+        schedule_dialing_tasks
+        # search peers every 15 seconds
+        task.reactor.every(15) do
+          schedule_dialing_tasks
         end
       end
 
+      private
+
+      def schedule_dialing_tasks
+        return unless @running_dialing < MAX_ACTIVE_DIAL_TASKS
+        @running_dialing += 1
+        @discovery_service.find_outgoing_peers(@running_dialing, @network_state.peers, Time.now).each do |node|
+          conn, handshake = @dialer.dial(node)
+          @network_state.new_peer_connected(conn, handshake)
+        end
+        @running_dialing -= 1
+      end
     end
+
   end
 end
+
