+++ lib/ciri/p2p/node.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-
 # Copyright (c) 2018 by Jiang Jinyang <jjyruby@gmail.com>
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -22,14 +21,65 @@
 # THE SOFTWARE.
 
 
-module Ciri
-  module DevP2P
+require 'ciri/key'
+require 'ciri/utils'
 
-    class Error < StandardError; end
-    class UselessPeerError < Error; end
-    class DisconnectError < Error; end
-    class UnknownMessageCodeError < Error; end
-    class InvalidMessageError < Error; end
+module Ciri
+  module P2P
+
+    # present node id
+    class NodeID
+
+      class << self
+        def from_raw_id(raw_id)
+          NodeID.new(Ciri::Key.new(raw_public_key: "\x04".b + raw_id))
+        end
+      end
+
+      attr_reader :public_key
+
+      alias key public_key
+
+      def initialize(public_key)
+        unless public_key.is_a?(Ciri::Key)
+          raise TypeError.new("expect Ciri::Key but get #{public_key.class}")
+        end
+        @public_key = public_key
+      end
+
+      def id
+        @id ||= key.raw_public_key[1..-1]
+      end
+
+      alias to_bytes id
+
+      def == (other)
+        self.class == other.class && id == other.id
+      end
+
+      def to_hex
+        Ciri::Utils.to_hex id
+      end
+
+      alias to_s to_hex
+
+    end
+
+    class Node
+      attr_reader :node_id, :ip, :udp_port, :tcp_port, :added_at
+
+      def initialize(node_id:, ip:, udp_port:, tcp_port:, added_at: nil)
+        @node_id = node_id
+        @ip = ip
+        @udp_port = udp_port
+        @tcp_port = tcp_port
+        @added_at = added_at
+      end
+
+      def == (other)
+        self.class == other.class && node_id == other.node_id
+      end
+    end
 
   end
 end
