+++ lib/ciri/p2p/rlpx/message.rb
@@ -21,66 +21,25 @@
 # THE SOFTWARE.
 
 
-require 'ciri/key'
-require 'ciri/utils'
+require 'ciri/rlp'
 
 module Ciri
-  module DevP2P
+  module P2P
+    module RLPX
 
-    # present node id
-    class NodeID
+      # RLPX message
+      class Message
+        include Ciri::RLP::Serializable
 
-      class << self
-        def from_raw_id(raw_id)
-          NodeID.new(Ciri::Key.new(raw_public_key: "\x04".b + raw_id))
-        end
+        attr_accessor :received_at
+
+        schema(
+            code: Integer,
+            size: Integer,
+            payload: RLP::Bytes
+        )
       end
-
-      attr_reader :public_key
-
-      alias key public_key
-
-      def initialize(public_key)
-        unless public_key.is_a?(Ciri::Key)
-          raise TypeError.new("expect Ciri::Key but get #{public_key.class}")
-        end
-        @public_key = public_key
-      end
-
-      def id
-        @id ||= key.raw_public_key[1..-1]
-      end
-
-      alias to_bytes id
-
-      def == (other)
-        self.class == other.class && id == other.id
-      end
-
-      def to_hex
-        Ciri::Utils.to_hex id
-      end
-
-      alias to_s to_hex
 
     end
-
-    class Node
-      attr_reader :node_id, :ip, :udp_port, :tcp_port, :added_at
-
-      def initialize(node_id:, ip:, udp_port:, tcp_port:, added_at: nil)
-        @node_id = node_id
-        @ip = ip
-        @udp_port = udp_port
-        @tcp_port = tcp_port
-        @added_at = added_at
-      end
-
-      def == (other)
-        self.class == other.class && node_id == other.node_id
-      end
-    end
-
   end
 end
-
